최대공약수 최대공배수 자료형 long으로 바꿈

+++ ex01/src/ex0316/최대공약수_최소공배수.java
@@ -5,11 +5,11 @@ public class 최대공약수_최소공배수 {
 	public static void main(String[] args) {
 		Scanner in = new Scanner(System.in);
 		System.out.print("첫 번째 정수를 입력하세요: ");
-		int a = in.nextInt();
+		long a = in.nextLong();
 		System.out.print("두 번째 정수를 입력하세요: ");
-		int b = in.nextInt();
-		int big = 0;
-		int small = 0;
+		long b = in.nextLong();
+		long big = 0;
+		long small = 0;
 		if(a<b) {
 			small = a;
 			big = b;
@@ -18,9 +18,9 @@ public class 최대공약수_최소공배수 {
 			big = a;
 			small= b;	
 		}
-		int gcm = 0;
+		long gcm = 0;
 		long lcm = 0;
-		int nmg = 0;
+		long nmg = 0;
 		while(true) {
 			
 			nmg = big%small;
